1981.2: Move types into `types.ts`
